src/main/python_macOs/requirements.txt
APScheduler==3.6.3
asgiref==3.5.2
branca==0.5.0
cachetools==4.2.2
certifi==2022.9.24
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.6
cycler==0.11.0
Django==4.1.2
et-xmlfile==1.1.0
fonttools==4.38.0
googletrans==3.1.0a0
greenlet==1.1.3.post0
gTTS==2.3.0
gTTS-token==1.1.4
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2022.10.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
image==1.5.33
imutils==0.5.4
Jinja2==3.1.2
kiwisolver==1.4.4
llvmlite==0.39.1
MarkupSafe==2.1.1
matplotlib==3.6.1
numpy==1.23.4
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
openpyxl==3.0.10
packaging==21.3
pandas==1.5.2
Pillow==9.2.0
playsound==1.2.2
psycopg2-binary==2.9.5
pycodestyle==2.9.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-telegram-bot==13.14
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
pyobject==1.2.1
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.42
sqlparse==0.4.3
tomli==2.0.1
telegram
tornado==6.1
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.12
wheel==0.38.4
xlrd==2.0.1
XlsxWriter==3.0.3
